Bayesian Treatment Effect Weight Estimation for Censored Data

This function estimates Bayesian importance sampling weights for treatment models and censoring models across multiple time points. The models can be run in parallel to estimate the weights for censored data.

Usage

bayesweight_cen(
  trtmodel.list,
  cenmodel.list,
  data,
  n.chains = 2,
  n.iter = 25000,
  n.burnin = 15000,
  n.thin = 5,
  seed = NULL,
  parallel = TRUE
)

Arguments

trtmodel.list

A list of formulas corresponding to each time point with the time-specific treatment variable on the left-hand side and pre-treatment covariates to be balanced on the right-hand side. The formulas must be in temporal order, and must contain all covariates to be balanced at that time point. Interactions and functions of covariates are allowed.

cenmodel.list

A list of formulas for the censored data at each time point, with censoring indicators on the left-hand side and covariates on the right-hand side. The formulas must be in temporal order, and must contain all covariates to be balanced at that time point.

data

A data set in the form of a data frame containing the variables in `trtmodel.list` and `cenmodel.list`. This must be a wide data set with exactly one row per unit.

n.chains

Integer specifying the number of MCMC chains to run. Set to 1 for non-parallel computation. For parallel computation, it is required to use at least 2 chains. The default is 2.

n.iter

Integer specifying the total number of iterations for each chain (including burn-in). The default is 25000.

n.burnin

Integer specifying the number of burn-in iterations for each chain. The default is 15000.

n.thin

Integer specifying the thinning rate for the MCMC sampler. The default is 5.

seed

Starting seed for the JAGS model. The default is NULL.

parallel

Logical scalar indicating whether to run the MCMC chains in parallel. The default is TRUE.

Value

A list of the calculated weights and the JAGS model: * `weights` is a vector of posterior mean weights, computed by taking the average of the weights across all MCMC iterations. * `model_string` is a character of the JAGS model based on the input of `trtmodel.list`.

Examples

simdat_cen <- read.csv(system.file("extdata",
                                   "sim_causal.csv",
                                   package = "bayesmsm"))
weights_cen <- bayesweight_cen(trtmodel.list = list(A1 ~ L11 + L21,
                                                    A2 ~ L11 + L21 + L12 +
                                                         L22 + A1,
                                                    A3 ~ L11 + L21 + L12 +
                                                         L22 + A1 + L13 +
                                                         L23 + A2),
                               cenmodel.list = list(C1 ~ L11 + L21,
                                                    C2 ~ L11 + L21 + A1,
                                                    C3 ~ L11 + L21 + A1 +
                                                         L12 + L22 + A2),
                               data = simdat_cen,
                               n.chains = 1,
                               n.iter = 200,
                               n.burnin = 100,
                               n.thin = 1,
                               seed = 890123,
                               parallel = FALSE)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 4836
#>    Unobserved stochastic nodes: 44
#>    Total graph size: 16501
#> 
#> Initializing model
#> 
#> Warning: Some parameters have not converged with Geweke index > 1.96. More iterations may be needed.
weights_cen$weights
#>   [1]         NA         NA  1.3181537  0.6485255  4.1984446  4.2858047
#>   [7]  2.4057620  1.1096888         NA  0.9239421  0.8402782         NA
#>  [13]  0.8262908         NA 21.8807863         NA  2.2559182  3.0520633
#>  [19]         NA  0.9110585         NA         NA  8.8497556  0.6613183
#>  [25]  0.5522206         NA  1.7377072  0.5366206  4.7858593         NA
#>  [31]  0.8305732  0.4478124         NA  0.4163883         NA  4.5067340
#>  [37]  2.8071403  4.0139072         NA         NA         NA  3.7833081
#>  [43]         NA         NA  2.3212080         NA  4.6130582         NA
#>  [49]  5.2636897         NA         NA  0.7579971  2.7346251         NA
#>  [55]  6.3765585  5.1705528  2.1971006  2.0538602  1.9112207         NA
#>  [61]  1.4961053  1.7823079  4.9105289  3.2669707  1.2347478  1.8808774
#>  [67]  0.9111320         NA  4.9034164  2.4281209  0.7911967 13.3051676
#>  [73]  0.5354100  3.7814541  4.5754552  1.9391805  1.6160621  1.6507833
#>  [79]  1.3579748 11.0605413 11.2496701  2.0134919         NA  0.8679503
#>  [85]  4.4257430  2.6819463         NA  4.3587090  3.1657073  0.9642465
#>  [91]  5.2180181  0.8183414         NA  0.6600481  2.2327731  2.6197462
#>  [97]  2.1153669  1.0602776  1.8076925  0.6557731  1.4692285  1.1231202
#> [103]  5.0492464  2.0519455  4.9522685  2.7935375  0.9832600  0.8377470
#> [109]  3.5968062  3.0923311  1.8434963  4.8913728         NA         NA
#> [115]         NA  1.3296895  1.7241407  0.6715433  3.8755273         NA
#> [121]  4.1412531  7.1872209  0.9587671  1.8816357  2.2843816  1.8421776
#> [127]  8.0558313         NA  2.4658013         NA  0.8535703  0.8178871
#> [133]  0.8173996  0.6502359  1.4936320  2.3071643  0.8727078         NA
#> [139]  4.4212193  1.7896425  2.2034573         NA  8.4839566         NA
#> [145]         NA         NA         NA  2.6885576  4.3228203  1.5628181
#> [151]  2.4898165         NA         NA  2.8605127         NA  0.8077778
#> [157]  0.6141652  1.4507573  1.3700963  4.7979925  4.4074743  1.5114750
#> [163]  5.0985477  3.7344909  4.1124723  3.1625812         NA  0.7519441
#> [169]  1.5927588  1.7599288  0.6916083  6.7100125  0.7401560  1.5215967
#> [175]  1.9871608  1.6413083 16.2795643  1.3206485  1.2812873  1.8887033
#> [181]  2.3788383  3.4381271         NA         NA         NA         NA
#> [187]  0.7861836         NA         NA         NA  8.0610921  3.2196394
#> [193]  4.9362088  3.1297085  2.7201227         NA         NA  2.6547595
#> [199]  1.1972010         NA         NA  1.9016187  1.3763587         NA
#> [205]  5.7431532         NA  2.7970490  5.6544056         NA  2.2397108
#> [211]         NA 10.4308607  2.3437941  2.7407671  0.8863485         NA
#> [217]  9.9501328  1.8133566         NA  0.8374330         NA  0.7408837
#> [223]  1.1368031         NA  1.3662972         NA  6.6666586  0.6271355
#> [229]  2.0480327  1.1115484         NA  4.0566075         NA         NA
#> [235]  7.9924678  2.5502084  1.2181483  0.7985679         NA  2.0410340
#> [241]  0.8393464         NA  7.0189784  0.7561783         NA  1.1766319
#> [247]  2.8183437  2.9884731         NA  0.5059034         NA  2.9670290
#> [253]  0.8844609  2.4119083         NA  0.8819635         NA         NA
#> [259] 11.3394804         NA  1.7571948  4.8550272  1.3629464  1.0110215
#> [265]  1.9953890         NA         NA  0.8969599  1.1299970         NA
#> [271]         NA  3.5829065  3.5321507  1.2755076  2.5395497  2.6073788
#> [277]  4.9263451  5.6455220  5.2720145  0.6002035  0.6627290  3.0202757
#> [283]  1.5793558         NA  2.4917939  3.7384110         NA 12.5676346
#> [289]  3.3249427         NA  1.4508147  4.6290895  1.7789868  0.8089185
#> [295]  2.0084167         NA  0.5247868         NA  7.3775327         NA
#> [301]  1.0080918  6.3983055  1.8614595  4.9171171         NA  1.5344659
#> [307]         NA         NA  2.3972883         NA         NA  3.2818032
#> [313]  1.7453256  3.4579307         NA  1.0935996  0.9354720         NA
#> [319]  0.8930711  0.8204181         NA  1.2309496  0.9912740  1.5108349
#> [325]  1.5822175  4.6665885  1.6093908  2.1227874  1.2721539  0.7234652
#> [331]  0.6376950  1.0582765         NA  3.0555811         NA  0.7964952
#> [337]  0.6921915  1.0112246         NA  0.8365573  6.6470461  8.5020017
#> [343]  2.1748780  3.5032713  4.5119748  1.1814941  0.6947016  1.0042193
#> [349]  8.6937670  2.8426572         NA  1.5480820  1.3672409  3.3675145
#> [355]  1.1868718  0.6299586         NA         NA  0.9015471         NA
#> [361]  4.4773897  1.4349205  0.5817465  3.5610906  8.4897917  0.3965976
#> [367]  1.2236736         NA  3.4381832  6.7269185  1.7589784  1.7189978
#> [373]  1.2883925 11.0814479  1.0939628         NA  2.2200863  1.6707923
#> [379]         NA         NA         NA  0.6993143         NA  1.7034153
#> [385]  1.1989015         NA  6.0349054  1.5312946  8.5835117  1.4042855
#> [391]  0.7249956         NA  0.5871502  0.8586319         NA  0.8228823
#> [397]  1.4790790  2.4312573  1.4139149  2.8310137         NA  4.0277707
#> [403]  0.9630813         NA  2.2940183  1.9494436  0.6949307  0.6841297
#> [409]  0.9053807  1.2735262  3.4251007  4.9185795  2.8324145  1.4420590
#> [415]         NA  3.7124241  0.6246150  1.2883962  3.5688647         NA
#> [421]  0.9338434  1.2767401         NA  1.0375994  0.5929075         NA
#> [427]  1.5311232         NA  1.7004277         NA  4.8470275  1.1154586
#> [433]         NA  1.6394186  1.1007082  0.6974090         NA         NA
#> [439]  1.2201631  4.0477722         NA  1.2658091  1.1117622  1.7594298
#> [445]  0.5902012         NA  0.8188042  1.3528233  7.0384224  0.6438076
#> [451]  4.9100404  1.3306198  1.4066671         NA  7.0061699  1.3449183
#> [457]  5.5754761  1.0825522  1.0713907  5.8950952         NA  5.5207962
#> [463]  4.2670232  3.9994025  2.0151120  6.5227921  6.7690953  0.7745359
#> [469]         NA         NA  0.7442837         NA         NA  0.5636045
#> [475]  0.6480853  2.5413645         NA  2.2138153         NA  1.0265385
#> [481]  2.6683794  0.8141443  5.3226586  1.1916202  2.8922595         NA
#> [487]         NA  1.2533536  1.9736422  2.1657345  1.1862123         NA
#> [493] 12.4666360  1.4691172  1.1339184         NA         NA  1.7317881
#> [499]  8.2914919  1.0191915
cat(weights_cen$model_string)
#> model{
#> 
#> for (i in 1:N1) {
#> 
#> # conditional model;
#> A1[i] ~ dbern(p1[i])
#> logit(p1[i]) <- b10 + b11*L11[i] + b12*L21[i]
#> C1[i] ~ dbern(cp1[i])
#> logit(cp1[i]) <- s10 + s11*L11[i] + s12*L21[i]
#> 
#> # marginal model;
#> A1s[i] ~ dbern(p1s[i])
#> logit(p1s[i]) <- bs10
#> C1s[i] ~ dbern(cp1s[i])
#> logit(cp1s[i]) <- ts10
#> }
#> 
#> for (i in 1:N2) {
#> 
#> # conditional model;
#> A2[i] ~ dbern(p2[i])
#> logit(p2[i]) <- b20 + b21*L11[i] + b22*L21[i] + b23*L12[i] + b24*L22[i] + b25*A1[i]
#> C2[i] ~ dbern(cp2[i])
#> logit(cp2[i]) <- s20 + s21*L11[i] + s22*L21[i] + s23*A1[i]
#> 
#> # marginal model;
#> A2s[i] ~ dbern(p2s[i])
#> logit(p2s[i]) <- bs20 + bs21*A1s[i]
#> C2s[i] ~ dbern(cp2s[i])
#> logit(cp2s[i]) <- ts20 + ts21*A1s[i]
#> }
#> 
#> for (i in 1:N3) {
#> 
#> # conditional model;
#> A3[i] ~ dbern(p3[i])
#> logit(p3[i]) <- b30 + b31*L11[i] + b32*L21[i] + b33*L12[i] + b34*L22[i] + b35*A1[i] + b36*L13[i] + b37*L23[i] + b38*A2[i]
#> C3[i] ~ dbern(cp3[i])
#> logit(cp3[i]) <- s30 + s31*L11[i] + s32*L21[i] + s33*A1[i] + s34*L12[i] + s35*L22[i] + s36*A2[i]
#> 
#> # marginal model;
#> A3s[i] ~ dbern(p3s[i])
#> logit(p3s[i]) <- bs30 + bs31*A1s[i] + bs32*A2s[i]
#> C3s[i] ~ dbern(cp3s[i])
#> logit(cp3s[i]) <- ts30 + ts31*A1s[i] + ts32*A2s[i]
#> }
#> 
#> # Priors
#> b10 ~ dunif(-10, 10)
#> b11 ~ dunif(-10, 10)
#> b12 ~ dunif(-10, 10)
#> s10 ~ dunif(-10, 10)
#> s11 ~ dunif(-10, 10)
#> s12 ~ dunif(-10, 10)
#> bs10 ~ dunif(-10, 10)
#> ts10 ~ dunif(-10, 10)
#> b20 ~ dunif(-10, 10)
#> b21 ~ dunif(-10, 10)
#> b22 ~ dunif(-10, 10)
#> b23 ~ dunif(-10, 10)
#> b24 ~ dunif(-10, 10)
#> b25 ~ dunif(-10, 10)
#> s20 ~ dunif(-10, 10)
#> s21 ~ dunif(-10, 10)
#> s22 ~ dunif(-10, 10)
#> s23 ~ dunif(-10, 10)
#> bs20 ~ dunif(-10, 10)
#> bs21 ~ dunif(-10, 10)
#> ts20 ~ dunif(-10, 10)
#> ts21 ~ dunif(-10, 10)
#> b30 ~ dunif(-10, 10)
#> b31 ~ dunif(-10, 10)
#> b32 ~ dunif(-10, 10)
#> b33 ~ dunif(-10, 10)
#> b34 ~ dunif(-10, 10)
#> b35 ~ dunif(-10, 10)
#> b36 ~ dunif(-10, 10)
#> b37 ~ dunif(-10, 10)
#> b38 ~ dunif(-10, 10)
#> s30 ~ dunif(-10, 10)
#> s31 ~ dunif(-10, 10)
#> s32 ~ dunif(-10, 10)
#> s33 ~ dunif(-10, 10)
#> s34 ~ dunif(-10, 10)
#> s35 ~ dunif(-10, 10)
#> s36 ~ dunif(-10, 10)
#> bs30 ~ dunif(-10, 10)
#> bs31 ~ dunif(-10, 10)
#> bs32 ~ dunif(-10, 10)
#> ts30 ~ dunif(-10, 10)
#> ts31 ~ dunif(-10, 10)
#> ts32 ~ dunif(-10, 10)
#> }