Bayesian Weight Estimation for Censored Data

This function computes posterior mean weights using Bayesian estimation for treatment models and censoring models across multiple time points. The models can be run in parallel to estimate the weights needed for causal inference with censored data.

Usage

bayesweight_cen(
  trtmodel.list = list(A1 ~ L11 + L21, A2 ~ L11 + L21 + L12 + L22 + A1, A3 ~ L11 + L21 +
    L12 + L22 + A1 + L13 + L23 + A2),
  cenmodel.list = list(C1 ~ L11 + L21, C2 ~ L11 + L21 + A1, C3 ~ L11 + L21 + A1 + L12 +
    L22 + A2),
  data,
  n.iter = 2500,
  n.burnin = 1500,
  n.thin = 5,
  parallel = FALSE,
  n.chains = 1,
  seed = 890123
)

Arguments

trtmodel.list

A list of formulas corresponding to each time point with the time-specific treatment variable on the left-hand side and pre-treatment covariates to be balanced on the right-hand side. Interactions and functions of covariates are allowed.

cenmodel.list

A list of formulas for the censoring data at each time point, with censoring indicators on the left-hand side and covariates on the right-hand side.

data

A data frame containing the variables in the models (treatment, censoring, and covariates).

n.iter

Number of iterations to run the MCMC algorithm in JAGS.

n.burnin

Number of iterations to discard as burn-in in the MCMC algorithm.

n.thin

Thinning rate for the MCMC samples.

parallel

Logical, indicating whether to run the MCMC sampling in parallel (default is `FALSE`).

n.chains

Number of MCMC chains to run. If parallel is `TRUE`, this specifies the number of chains run in parallel.

seed

A seed for random number generation to ensure reproducibility of the MCMC.

Value

A vector of posterior mean weights, computed by taking the average of the weights across all MCMC iterations.

Examples

simdat_cen <- read.csv(system.file("extdata",
                                   "sim_causal.csv",
                                   package = "bayesmsm"))
weights_cen <- bayesweight_cen(trtmodel.list = list(A1 ~ L11 + L21,
                                                    A2 ~ L11 + L21 + L12 +
                                                         L22 + A1,
                                                    A3 ~ L11 + L21 + L12 +
                                                         L22 + A1 + L13 +
                                                         L23 + A2),
                               cenmodel.list = list(C1 ~ L11 + L21,
                                                    C2 ~ L11 + L21 + A1,
                                                    C3 ~ L11 + L21 + A1 +
                                                         L12 + L22 + A2),
                               data = simdat_cen,
                               n.iter = 2500,
                               n.burnin = 1500,
                               n.thin = 5,
                               parallel = FALSE,
                               n.chains = 1,
                               seed = 890123)
#> module glm loaded
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 4836
#>    Unobserved stochastic nodes: 44
#>    Total graph size: 16501
#> 
#> Initializing model
#> 
#> Warning: Some parameters have not converged with Geweke index > 1.96. More iterations may be needed.