Bayesian Treatment Effect Weight Estimation for Censored Data

This function estimates Bayesian importance sampling weights for treatment models and censoring models across multiple time points. The models can be run in parallel to estimate the weights for censored data.

Usage

bayesweight_cen(
  trtmodel.list,
  cenmodel.list,
  data,
  n.chains = 2,
  n.iter = 25000,
  n.burnin = 15000,
  n.thin = 5,
  seed = NULL,
  parallel = TRUE
)

Arguments

trtmodel.list

A list of formulas corresponding to each time point with the time-specific treatment variable on the left-hand side and pre-treatment covariates to be balanced on the right-hand side. The formulas must be in temporal order, and must contain all covariates to be balanced at that time point. Interactions and functions of covariates are allowed.

cenmodel.list

A list of formulas for the censored data at each time point, with censoring indicators on the left-hand side and covariates on the right-hand side. The formulas must be in temporal order, and must contain all covariates to be balanced at that time point.

data

A data set in the form of a data frame containing the variables in `trtmodel.list` and `cenmodel.list`. This must be a wide data set with exactly one row per unit.

n.chains

Integer specifying the number of MCMC chains to run. Set to 1 for non-parallel computation. For parallel computation, it is required to use at least 2 chains. The default is 2.

n.iter

Integer specifying the total number of iterations for each chain (including burn-in). The default is 25000.

n.burnin

Integer specifying the number of burn-in iterations for each chain. The default is 15000.

n.thin

Integer specifying the thinning rate for the MCMC sampler. The default is 5.

seed

Starting seed for the JAGS model. The default is NULL.

parallel

Logical scalar indicating whether to run the MCMC chains in parallel. The default is TRUE.

Value

A list of the calculated weights and the JAGS model: * `weights` is a vector of posterior mean weights, computed by taking the average of the weights across all MCMC iterations. * `model_string` is a character of the JAGS model based on the input of `trtmodel.list`.

Examples

simdat_cen <- read.csv(system.file("extdata",
                                   "sim_causal.csv",
                                   package = "bayesmsm"))
weights_cen <- bayesweight_cen(trtmodel.list = list(A1 ~ L11 + L21,
                                                    A2 ~ L11 + L21 + L12 +
                                                         L22 + A1,
                                                    A3 ~ L11 + L21 + L12 +
                                                         L22 + A1 + L13 +
                                                         L23 + A2),
                               cenmodel.list = list(C1 ~ L11 + L21,
                                                    C2 ~ L11 + L21 + A1,
                                                    C3 ~ L11 + L21 + A1 +
                                                         L12 + L22 + A2),
                               data = simdat_cen,
                               n.chains = 1,
                               n.iter = 100,
                               n.burnin = 50,
                               n.thin = 1,
                               seed = 890123,
                               parallel = FALSE)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 4836
#>    Unobserved stochastic nodes: 44
#>    Total graph size: 16501
#> 
#> Initializing model
#> 
#> Warning: Some parameters have not converged with Geweke index > 1.96. More iterations may be needed.
weights_cen$weights
#>   [1]         NA         NA  1.2642557  0.5421839  5.1068112  4.6106131
#>   [7]  2.5628611  1.1624346         NA  0.8615265  0.6453359         NA
#>  [13]  0.7702220         NA 41.9686432         NA  2.1124888  2.8310504
#>  [19]         NA  0.8599781         NA         NA 11.0528462  0.6037800
#>  [25]  0.4256688         NA  1.6986594  0.4565428  4.0857228         NA
#>  [31]  0.6649828  0.4458072         NA  0.3572363         NA  4.7989275
#>  [37]  2.7796383  5.3666891         NA         NA         NA  3.1877134
#>  [43]         NA         NA  2.4763515         NA  5.7293331         NA
#>  [49]  4.4730435         NA         NA  0.6670697  1.8182251         NA
#>  [55]  6.1133190  5.1002404  2.4058608  1.6076956  2.0991117         NA
#>  [61]  1.6207057  1.8203088  4.0767679  2.7909660  0.9954311  2.2350067
#>  [67]  0.8012886         NA  5.3213842  2.3524277  0.6137062 14.4063376
#>  [73]  0.4814783  3.0585212  3.6917356  1.8514919  1.2693188  2.0618678
#>  [79]  1.2461243 10.9085170 10.9433822  1.4853800         NA  0.7802038
#>  [85]  3.7954172  2.9981071         NA  4.3043149  5.0666072  0.7836558
#>  [91]  5.5972972  0.7154000         NA  0.7297750  2.2584313  2.0463820
#>  [97]  1.8435590  0.9200094  1.6965353  0.5025329  1.1593356  1.1835386
#> [103]  5.8004972  1.5258320  4.5279686  2.5967291  1.0488590  0.8370011
#> [109]  3.2983603  3.1964225  2.2954880  4.2662150         NA         NA
#> [115]         NA  1.3024985  1.4539403  0.4789716  4.7365412         NA
#> [121]  9.2243106  6.8533719  0.8516919  2.1079709  2.1741208  1.6865729
#> [127]  8.6348186         NA  2.4642895         NA  0.7949080  0.6612319
#> [133]  0.7632680  0.4906116  1.9558100  3.8315631  1.0480859         NA
#> [139]  4.0974392  1.8327631  1.6497830         NA  8.8175462         NA
#> [145]         NA         NA         NA  3.0783729  3.5140238  1.8216870
#> [151]  2.5514611         NA         NA  3.0596808         NA  0.5479098
#> [157]  0.6114283  1.4002235  1.1771243  6.1854007  7.6027620  1.2965336
#> [163]  4.6876578  3.3932709  3.6855641  2.3783546         NA  0.6948663
#> [169]  1.5658269  1.7966367  0.6553735  6.3626982  0.5978526  1.5615526
#> [175]  1.5751493  1.7097782 25.2752348  1.4814711  1.3781355  1.1520594
#> [181]  1.9871549  3.6455640         NA         NA         NA         NA
#> [187]  0.7670094         NA         NA         NA  7.5108366  2.9045571
#> [193]  5.8053818  2.1969254  1.9318729         NA         NA  2.6854334
#> [199]  0.9927477         NA         NA  2.6585032  1.2317302         NA
#> [205]  7.1478271         NA  2.9214279  4.4036496         NA  2.4948547
#> [211]         NA 11.1668383  2.4768534  2.8077242  0.6316468         NA
#> [217] 10.5717261  1.8439924         NA  0.9024179         NA  0.6123551
#> [223]  0.9126017         NA  1.1493831         NA  7.0997746  0.4974687
#> [229]  1.8109279  1.0773890         NA  3.9436728         NA         NA
#> [235]  9.2009156  2.2599553  1.1410514  0.9189370         NA  2.6868934
#> [241]  0.6890612         NA  8.0389762  0.6135201         NA  0.9669870
#> [247]  3.0931571  2.7847929         NA  0.4276466         NA  2.9572903
#> [253]  0.8167080  2.5929929         NA  0.7530578         NA         NA
#> [259] 13.3556901         NA  1.4256164  4.7082256  1.5310912  1.0479520
#> [265]  1.5852587         NA         NA  0.8977643  0.9746008         NA
#> [271]         NA  3.8508909  3.8901540  1.1899726  2.2713094  1.9900367
#> [277]  5.1859714  4.5829754  8.1537514  0.4611933  0.5965306  2.6109000
#> [283]  1.5281177         NA  2.9949116  3.4104411         NA 10.6960381
#> [289]  3.2739967         NA  1.4204256  4.3792715  1.3842232  0.6353565
#> [295]  2.7143352         NA  0.4107999         NA  8.5216926         NA
#> [301]  1.0437557  4.7639776  2.2734899  5.4781574         NA  1.4494831
#> [307]         NA         NA  2.6064797         NA         NA  2.9247011
#> [313]  1.7279637  3.7861368         NA  1.1352616  0.8403243         NA
#> [319]  0.7090024  0.7228321         NA  1.0829929  0.8645996  1.4862268
#> [325]  1.4680649  3.7882913  1.6797210  1.6397124  1.0821975  0.6213965
#> [331]  0.4988684  0.9434763         NA  2.4011062         NA  0.6251752
#> [337]  0.5883513  1.0465093         NA  0.7902789  6.4139946  7.8854022
#> [343]  1.7026138  2.9833726  3.7689096  0.9666110  0.5736311  0.8611906
#> [349] 10.2891220  3.8340888         NA  1.7631570  1.2640729  2.8286075
#> [355]  1.3177636  0.5049690         NA         NA  0.8141416         NA
#> [361]  3.5503212  1.4486297  0.5244178  3.1316260  9.9784253  0.3168891
#> [367]  1.0685940         NA  2.7226870  8.1026539  1.6659846  1.7689128
#> [373]  1.4354677 10.6149158  1.0297974         NA  1.9091373  1.3922008
#> [379]         NA         NA         NA  0.5082350         NA  1.9850473
#> [385]  0.7465171         NA  4.7496394  1.5066564  7.6490448  1.0315575
#> [391]  0.7329128         NA  0.4396534  0.8033730         NA  0.7247949
#> [397]  1.3065018  2.2666238  1.3333333  3.0364208         NA  4.0021068
#> [403]  1.0096923         NA  2.4445359  1.7927218  0.5483263  0.6158796
#> [409]  0.8165261  1.3123588  3.2731828  5.1866377  2.7314379  1.3067124
#> [415]         NA  3.3024311  0.5434301  1.1772067  3.5149072         NA
#> [421]  0.8110116  0.9197182         NA  0.8483333  0.4712121         NA
#> [427]  1.3315163         NA  1.7809490         NA  5.0221243  1.1214163
#> [433]         NA  1.8635812  0.9970177  0.5173897         NA         NA
#> [439]  1.1142728  3.5302873         NA  1.2665887  1.1950861  1.6939471
#> [445]  0.5218719         NA  0.7313385  1.1894313  6.7359700  0.6052058
#> [451]  6.0693534  1.1289624  1.2707285         NA  6.3298339  1.4326902
#> [457]  5.3628578  0.8424320  0.8937966  5.6160488         NA  4.4530376
#> [463]  3.6230589  4.5539411  2.7234565  6.2563164  7.6627906  0.6145252
#> [469]         NA         NA  0.7000282         NA         NA  0.3420752
#> [475]  0.5480696  2.8847894         NA  1.8835697         NA  0.8873318
#> [481]  2.6136825  0.7934250  5.1132869  1.4806361  2.7292906         NA
#> [487]         NA  1.1424842  1.8419137  1.7765518  1.2875978         NA
#> [493] 16.0166576  1.3178576  1.0443955         NA         NA  1.5957681
#> [499] 12.9504441  0.8241830
cat(weights_cen$model_string)
#> model{
#> 
#> for (i in 1:N1) {
#> 
#> # conditional model;
#> A1[i] ~ dbern(p1[i])
#> logit(p1[i]) <- b10 + b11*L11[i] + b12*L21[i]
#> C1[i] ~ dbern(cp1[i])
#> logit(cp1[i]) <- s10 + s11*L11[i] + s12*L21[i]
#> 
#> # marginal model;
#> A1s[i] ~ dbern(p1s[i])
#> logit(p1s[i]) <- bs10
#> C1s[i] ~ dbern(cp1s[i])
#> logit(cp1s[i]) <- ts10
#> }
#> 
#> for (i in 1:N2) {
#> 
#> # conditional model;
#> A2[i] ~ dbern(p2[i])
#> logit(p2[i]) <- b20 + b21*L11[i] + b22*L21[i] + b23*L12[i] + b24*L22[i] + b25*A1[i]
#> C2[i] ~ dbern(cp2[i])
#> logit(cp2[i]) <- s20 + s21*L11[i] + s22*L21[i] + s23*A1[i]
#> 
#> # marginal model;
#> A2s[i] ~ dbern(p2s[i])
#> logit(p2s[i]) <- bs20 + bs21*A1s[i]
#> C2s[i] ~ dbern(cp2s[i])
#> logit(cp2s[i]) <- ts20 + ts21*A1s[i]
#> }
#> 
#> for (i in 1:N3) {
#> 
#> # conditional model;
#> A3[i] ~ dbern(p3[i])
#> logit(p3[i]) <- b30 + b31*L11[i] + b32*L21[i] + b33*L12[i] + b34*L22[i] + b35*A1[i] + b36*L13[i] + b37*L23[i] + b38*A2[i]
#> C3[i] ~ dbern(cp3[i])
#> logit(cp3[i]) <- s30 + s31*L11[i] + s32*L21[i] + s33*A1[i] + s34*L12[i] + s35*L22[i] + s36*A2[i]
#> 
#> # marginal model;
#> A3s[i] ~ dbern(p3s[i])
#> logit(p3s[i]) <- bs30 + bs31*A1s[i] + bs32*A2s[i]
#> C3s[i] ~ dbern(cp3s[i])
#> logit(cp3s[i]) <- ts30 + ts31*A1s[i] + ts32*A2s[i]
#> }
#> 
#> # Priors
#> b10 ~ dunif(-10, 10)
#> b11 ~ dunif(-10, 10)
#> b12 ~ dunif(-10, 10)
#> s10 ~ dunif(-10, 10)
#> s11 ~ dunif(-10, 10)
#> s12 ~ dunif(-10, 10)
#> bs10 ~ dunif(-10, 10)
#> ts10 ~ dunif(-10, 10)
#> b20 ~ dunif(-10, 10)
#> b21 ~ dunif(-10, 10)
#> b22 ~ dunif(-10, 10)
#> b23 ~ dunif(-10, 10)
#> b24 ~ dunif(-10, 10)
#> b25 ~ dunif(-10, 10)
#> s20 ~ dunif(-10, 10)
#> s21 ~ dunif(-10, 10)
#> s22 ~ dunif(-10, 10)
#> s23 ~ dunif(-10, 10)
#> bs20 ~ dunif(-10, 10)
#> bs21 ~ dunif(-10, 10)
#> ts20 ~ dunif(-10, 10)
#> ts21 ~ dunif(-10, 10)
#> b30 ~ dunif(-10, 10)
#> b31 ~ dunif(-10, 10)
#> b32 ~ dunif(-10, 10)
#> b33 ~ dunif(-10, 10)
#> b34 ~ dunif(-10, 10)
#> b35 ~ dunif(-10, 10)
#> b36 ~ dunif(-10, 10)
#> b37 ~ dunif(-10, 10)
#> b38 ~ dunif(-10, 10)
#> s30 ~ dunif(-10, 10)
#> s31 ~ dunif(-10, 10)
#> s32 ~ dunif(-10, 10)
#> s33 ~ dunif(-10, 10)
#> s34 ~ dunif(-10, 10)
#> s35 ~ dunif(-10, 10)
#> s36 ~ dunif(-10, 10)
#> bs30 ~ dunif(-10, 10)
#> bs31 ~ dunif(-10, 10)
#> bs32 ~ dunif(-10, 10)
#> ts30 ~ dunif(-10, 10)
#> ts31 ~ dunif(-10, 10)
#> ts32 ~ dunif(-10, 10)
#> }