Bayesian Treatment Effect Weight Estimation for Censored Data

This function estimates Bayesian importance sampling weights for treatment models and censoring models across multiple time points. The models can be run in parallel to estimate the weights for censored data.

Usage

bayesweight_cen(
  trtmodel.list,
  cenmodel.list,
  data,
  n.chains = 2,
  n.iter = 25000,
  n.burnin = 15000,
  n.thin = 5,
  seed = NULL,
  parallel = TRUE
)

Arguments

trtmodel.list

A list of formulas corresponding to each time point with the time-specific treatment variable on the left-hand side and pre-treatment covariates to be balanced on the right-hand side. The formulas must be in temporal order, and must contain all covariates to be balanced at that time point. Interactions and functions of covariates are allowed.

cenmodel.list

A list of formulas for the censored data at each time point, with censoring indicators on the left-hand side and covariates on the right-hand side. The formulas must be in temporal order, and must contain all covariates to be balanced at that time point.

data

A data set in the form of a data frame containing the variables in `trtmodel.list` and `cenmodel.list`. This must be a wide data set with exactly one row per unit.

n.chains

Integer specifying the number of MCMC chains to run. Set to 1 for non-parallel computation. For parallel computation, it is required to use at least 2 chains. The default is 2.

n.iter

Integer specifying the total number of iterations for each chain (including burn-in). The default is 25000.

n.burnin

Integer specifying the number of burn-in iterations for each chain. The default is 15000.

n.thin

Integer specifying the thinning rate for the MCMC sampler. The default is 5.

seed

Starting seed for the JAGS model. The default is NULL.

parallel

Logical scalar indicating whether to run the MCMC chains in parallel. The default is TRUE.

Value

A list of the calculated weights and the JAGS model: * `weights` is a vector of posterior mean weights, computed by taking the average of the weights across all MCMC iterations. * `model_string` is a character of the JAGS model based on the input of `trtmodel.list`.

Examples

simdat_cen <- read.csv(system.file("extdata",
                                   "sim_causal.csv",
                                   package = "bayesmsm"))
weights_cen <- bayesweight_cen(trtmodel.list = list(A1 ~ L11 + L21,
                                                    A2 ~ L11 + L21 + L12 +
                                                         L22 + A1,
                                                    A3 ~ L11 + L21 + L12 +
                                                         L22 + A1 + L13 +
                                                         L23 + A2),
                               cenmodel.list = list(C1 ~ L11 + L21,
                                                    C2 ~ L11 + L21 + A1,
                                                    C3 ~ L11 + L21 + A1 +
                                                         L12 + L22 + A2),
                               data = simdat_cen,
                               n.chains = 1,
                               n.iter = 20,
                               n.burnin = 10,
                               n.thin = 1,
                               seed = 890123,
                               parallel = FALSE)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 4836
#>    Unobserved stochastic nodes: 44
#>    Total graph size: 16501
#> 
#> Initializing model
#> 
#> Warning: Some parameters have not converged with Geweke index > 1.96. More iterations may be needed.
weights_cen$weights
#>   [1]         NA         NA  1.4456133  1.0715142  6.3767871  0.9070324
#>   [7]  3.3210244  1.8321974         NA  0.5592724  0.2442320         NA
#>  [13]  1.4225262         NA 37.8490445         NA  0.4769442  3.0263779
#>  [19]         NA  1.5506225         NA         NA  3.7260737  0.7879442
#>  [25]  0.6917942         NA  0.4741509  0.4868215  1.3762581         NA
#>  [31]  1.7644462  0.4441765         NA  0.4472510         NA  5.1561126
#>  [37]  1.2810546  5.7495824         NA         NA         NA  1.0768918
#>  [43]         NA         NA  1.8367064         NA  5.6389345         NA
#>  [49]  1.4975100         NA         NA  0.4997192  1.1363787         NA
#>  [55]  2.8320738  1.7342932  0.4459201  0.5279491  0.5270837         NA
#>  [61]  2.0478475  1.9635426  1.2480141  1.0049444  2.0639244  0.4959758
#>  [67]  1.5214438         NA  1.7043326  1.7448061  1.1183049  4.4624114
#>  [73]  0.6512373  1.0563612  1.2400973  1.3442917  0.3411946  1.6669307
#>  [79]  0.9193874  3.3864862  3.5010020  1.0749686         NA  0.8050216
#>  [85]  1.2781515  0.5329488         NA  1.9485849  0.9821594  0.6236220
#>  [91]  1.8419728  1.0563245         NA  0.6967079  2.4293466  0.6809041
#>  [97]  0.9762709  1.0680183  1.9336844  1.0718710  0.6376153  0.6996744
#> [103]  6.2722220  0.7656690  1.5343489  1.2324013  0.8898772  0.8659817
#> [109]  0.9459425  2.7988221  0.5811587  1.4012873         NA         NA
#> [115]         NA  0.9036810  0.7869178  1.0150495  0.9925048         NA
#> [121]  1.3388787  2.1164757  0.6054858  0.4968597  1.6053360  2.7785617
#> [127]  1.4506135         NA  2.7189633         NA  1.4651409  1.3614218
#> [133]  0.7406890  1.2300459  0.4787525  0.7312590  0.9020800         NA
#> [139]  2.0741404  0.4468571  0.5854721         NA  1.8021092         NA
#> [145]         NA         NA         NA  0.7037651  1.2322136  0.5022056
#> [151]  3.7040171         NA         NA  2.0604004         NA  0.6819747
#> [157]  0.6494412  1.5645470  1.3689165  1.1875363  1.3182951  0.3329154
#> [163]  1.7523019  1.1968655  1.2928663  1.6249260         NA  0.5358260
#> [169]  1.7634216  2.0409005  0.9906185  2.0518486  1.2544008  2.0134260
#> [175]  1.0081397  3.1160102  6.1098047  0.4252421  1.0909677  0.8257719
#> [181]  1.0980287  3.9284680         NA         NA         NA         NA
#> [187]  0.5765073         NA         NA         NA  2.3200655  1.0379919
#> [193]  2.1571248  1.1832373  1.0393904         NA         NA  1.1582286
#> [199]  1.0410025         NA         NA  0.5988185  0.3856283         NA
#> [205]  1.3315718         NA  0.6538336  1.4771296         NA  1.6878495
#> [211]         NA  3.4668441  0.5897674  4.6215902  0.6943383         NA
#> [217]  3.4372221  1.9664473         NA  0.8104503         NA  1.2634746
#> [223]  0.3525822         NA  0.2904560         NA  1.5087182  1.0780067
#> [229]  0.9258581  1.1137677         NA  2.7953670         NA         NA
#> [235]  2.9118667  0.5668535  0.8308761  0.8993407         NA  0.6487051
#> [241]  1.4571363         NA  2.5559407  0.2356121         NA  0.3428758
#> [247]  0.6861395  1.3287800         NA  0.4308363         NA  0.6082375
#> [253]  0.8363835  1.8886447         NA  1.4499738         NA         NA
#> [259]  4.0740874         NA  0.8362394  2.1127773  2.2395512  1.0215676
#> [265]  0.9952046         NA         NA  0.6610745  1.1425156         NA
#> [271]         NA  4.1480810  4.0564595  0.8689779  0.9018808  0.9783292
#> [277]  2.0577342  1.5355985  1.3854943  1.0600955  1.0559984  0.9412578
#> [283]  1.7413813         NA  0.7103160  0.8068407         NA  3.4827274
#> [289]  0.6625502         NA  1.6204999  1.4945044  0.3961832  1.5388599
#> [295]  0.6248668         NA  0.5146343         NA  2.3326226         NA
#> [301]  1.6243442  1.6706995  3.0788630  2.1727795         NA  1.6599689
#> [307]         NA         NA  0.5672330         NA         NA  1.0512478
#> [313]  3.5016230  3.9159424         NA  0.3570301  1.2768014         NA
#> [319]  0.8274960  1.3580063         NA  1.2694614  1.0038738  0.4780249
#> [325]  1.5702158  1.2814873  1.5303948  0.6270647  1.2761410  1.2092664
#> [331]  1.0827696  1.0789058         NA  1.0627837         NA  0.7235668
#> [337]  1.1008483  0.3534820         NA  0.3149297  1.9165977  2.5913339
#> [343]  0.7730444  1.0780128  1.2814587  1.0350312  1.1674616  0.5640669
#> [349]  2.1148995  4.5271434         NA  0.4879293  1.9530330  1.0044210
#> [355]  1.9785729  1.0685356         NA         NA  0.5781106         NA
#> [361]  1.2009846  0.3138492  0.4777289  0.7180741  1.9054755  0.5487007
#> [367]  1.1429087         NA  0.9893906  2.5057329  0.3399657  0.4294825
#> [373]  1.2590526  3.3555596  2.1176526         NA  1.0037328  1.4862654
#> [379]         NA         NA         NA  1.5053632         NA  3.0063567
#> [385]  0.6905174         NA  1.5927066  1.6559885  2.4319522  0.2851862
#> [391]  0.8360234         NA  0.9172651  1.3624844         NA  0.5351303
#> [397]  1.5137546  0.5017157  1.4953819  1.2393174         NA  2.7721061
#> [403]  0.9551280         NA  2.7561145  2.0318988  0.9277557  0.4825569
#> [409]  0.5973345  2.1850863  3.5942345  1.7187104  0.6732510  1.3931529
#> [415]         NA  1.1775946  0.5614513  1.3514378  3.6329015         NA
#> [421]  1.6391181  0.6338095         NA  0.9956453  0.9630407         NA
#> [427]  1.4170987         NA  2.0208076         NA  2.0550856  1.7279942
#> [433]         NA  0.5467121  0.7137778  1.2377496         NA         NA
#> [439]  1.2810709  1.1962992         NA  1.7409836  1.7464953  1.9314869
#> [445]  0.7202878         NA  1.4186516  1.3744740  2.2063415  0.8127669
#> [451]  1.3975347  1.3293163  1.4700721         NA  2.0259989  0.4385801
#> [457]  2.3511732  1.9721742  0.7115987  1.2397149         NA  1.5004124
#> [463]  1.2777516  5.0681893  3.3907895  2.1549335  2.4538680  1.3728407
#> [469]         NA         NA  0.4996574         NA         NA  1.0470441
#> [475]  1.0915790  0.6321866         NA  0.4714820         NA  1.2945633
#> [481]  0.6393459  1.2854676  1.6224881  1.1690034  1.3333225         NA
#> [487]         NA  1.8120207  1.9643350  0.7704345  0.3947983         NA
#> [493]  4.2129698  0.3870741  1.1818599         NA         NA  1.8135385
#> [499]  3.1182565  1.6182261
cat(weights_cen$model_string)
#> model{
#> 
#> for (i in 1:N1) {
#> 
#> # conditional model;
#> A1[i] ~ dbern(p1[i])
#> logit(p1[i]) <- b10 + b11*L11[i] + b12*L21[i]
#> C1[i] ~ dbern(cp1[i])
#> logit(cp1[i]) <- s10 + s11*L11[i] + s12*L21[i]
#> 
#> # marginal model;
#> A1s[i] ~ dbern(p1s[i])
#> logit(p1s[i]) <- bs10
#> C1s[i] ~ dbern(cp1s[i])
#> logit(cp1s[i]) <- ts10
#> }
#> 
#> for (i in 1:N2) {
#> 
#> # conditional model;
#> A2[i] ~ dbern(p2[i])
#> logit(p2[i]) <- b20 + b21*L11[i] + b22*L21[i] + b23*L12[i] + b24*L22[i] + b25*A1[i]
#> C2[i] ~ dbern(cp2[i])
#> logit(cp2[i]) <- s20 + s21*L11[i] + s22*L21[i] + s23*A1[i]
#> 
#> # marginal model;
#> A2s[i] ~ dbern(p2s[i])
#> logit(p2s[i]) <- bs20 + bs21*A1s[i]
#> C2s[i] ~ dbern(cp2s[i])
#> logit(cp2s[i]) <- ts20 + ts21*A1s[i]
#> }
#> 
#> for (i in 1:N3) {
#> 
#> # conditional model;
#> A3[i] ~ dbern(p3[i])
#> logit(p3[i]) <- b30 + b31*L11[i] + b32*L21[i] + b33*L12[i] + b34*L22[i] + b35*A1[i] + b36*L13[i] + b37*L23[i] + b38*A2[i]
#> C3[i] ~ dbern(cp3[i])
#> logit(cp3[i]) <- s30 + s31*L11[i] + s32*L21[i] + s33*A1[i] + s34*L12[i] + s35*L22[i] + s36*A2[i]
#> 
#> # marginal model;
#> A3s[i] ~ dbern(p3s[i])
#> logit(p3s[i]) <- bs30 + bs31*A1s[i] + bs32*A2s[i]
#> C3s[i] ~ dbern(cp3s[i])
#> logit(cp3s[i]) <- ts30 + ts31*A1s[i] + ts32*A2s[i]
#> }
#> 
#> # Priors
#> b10 ~ dunif(-10, 10)
#> b11 ~ dunif(-10, 10)
#> b12 ~ dunif(-10, 10)
#> s10 ~ dunif(-10, 10)
#> s11 ~ dunif(-10, 10)
#> s12 ~ dunif(-10, 10)
#> bs10 ~ dunif(-10, 10)
#> ts10 ~ dunif(-10, 10)
#> b20 ~ dunif(-10, 10)
#> b21 ~ dunif(-10, 10)
#> b22 ~ dunif(-10, 10)
#> b23 ~ dunif(-10, 10)
#> b24 ~ dunif(-10, 10)
#> b25 ~ dunif(-10, 10)
#> s20 ~ dunif(-10, 10)
#> s21 ~ dunif(-10, 10)
#> s22 ~ dunif(-10, 10)
#> s23 ~ dunif(-10, 10)
#> bs20 ~ dunif(-10, 10)
#> bs21 ~ dunif(-10, 10)
#> ts20 ~ dunif(-10, 10)
#> ts21 ~ dunif(-10, 10)
#> b30 ~ dunif(-10, 10)
#> b31 ~ dunif(-10, 10)
#> b32 ~ dunif(-10, 10)
#> b33 ~ dunif(-10, 10)
#> b34 ~ dunif(-10, 10)
#> b35 ~ dunif(-10, 10)
#> b36 ~ dunif(-10, 10)
#> b37 ~ dunif(-10, 10)
#> b38 ~ dunif(-10, 10)
#> s30 ~ dunif(-10, 10)
#> s31 ~ dunif(-10, 10)
#> s32 ~ dunif(-10, 10)
#> s33 ~ dunif(-10, 10)
#> s34 ~ dunif(-10, 10)
#> s35 ~ dunif(-10, 10)
#> s36 ~ dunif(-10, 10)
#> bs30 ~ dunif(-10, 10)
#> bs31 ~ dunif(-10, 10)
#> bs32 ~ dunif(-10, 10)
#> ts30 ~ dunif(-10, 10)
#> ts31 ~ dunif(-10, 10)
#> ts32 ~ dunif(-10, 10)
#> }